+++ src/index.ts
@@ -14,7 +14,7 @@ import { Plugins } from './redteam/plugins';
 import { RedteamPluginBase, RedteamGraderBase } from './redteam/plugins/base';
 import { Strategies } from './redteam/strategies';
 import telemetry from './telemetry';
-import { readTests } from './testCases';
+import { readTests } from './util/testCaseReader';
 import type {
   EvaluateOptions,
   TestSuite,
